Move nb utilities
